tests/wpt/tests/html/canvas/offscreen/compositing/2d.composite.globalAlpha.canvas.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.composite.globalAlpha.canvas</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.composite.globalAlpha.canvas</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var offscreenCanvas2 = new OffscreenCanvas(100, 50);
  var ctx2 = offscreenCanvas2.getContext('2d');
  ctx2.fillStyle = '#f00';
  ctx2.fillRect(0, 0, 100, 50);
  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  // Avoiding any potential alpha = 0 optimisations.
  ctx.globalAlpha = 0.01;
  ctx.drawImage(offscreenCanvas2, 0, 0);
  _assertPixelApprox(canvas, 50,25, 2,253,0,255, 2);
  t.done();

});
</script>
